docs(tailwind): extract psuedo stories & add TW focus rings
