+++ packages/tailwind/docs/TailwindTokens.stories.tsx
@@ -56,34 +56,6 @@ export const DefaultKaizenSiteDemo = args => (
 DefaultKaizenSiteDemo.storyName = "Default (Kaizen Site Demo)"
 DefaultKaizenSiteDemo.args = { title: "Tailwind title example" }
 
-export const TailwindPsuedoStates = () => (
-  <div className="p-12">
-    <h1 className="flex flex-col items-center text-heading-1 font-weight-heading font-family-heading text-purple-700">
-      Tailwind Psuedo states
-    </h1>
-    <div className="flex">
-      <div className="inline-flex flex-col">
-        <h2 className="text-heading-2 font-family-heading-2 font-weight-heading-2 text-purple-800 mb-12">
-          Hover
-        </h2>
-        <button className=" rounded-default border-none p-12 bg-purple-600 text-white font-family-paragraph font-weight-paragraph-bold text-paragraph rounded border-width-default  hover:bg-purple-800">
-          Learn more
-        </button>
-      </div>
-      <div className="inline-flex flex-col ml-12">
-        <h2 className="text-heading-2 font-family-heading-2 font-weight-heading-2 text-purple-800 mb-12">
-          Focus
-        </h2>
-        <button className="border-none rounded-default p-12 bg-purple-600 text-white font-family-paragraph font-weight-paragraph-bold text-paragraph rounded focus:ring-2 focus:ring-offset-1 focus:bg-purple-800">
-          Learn more
-        </button>
-      </div>
-    </div>
-  </div>
-)
-
-TailwindPsuedoStates.storyName = "Tailwind psuedo selectors"
-
 export const TailwindMediaQueries = () => (
   <div className="p-12">
     <h1 className="flex flex-col items-center text-heading-1 font-weight-heading font-family-heading text-purple-700">
